% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_format.R
\name{print_trials_js}
\alias{print_trials_js}
\title{Print trial list as JavaScript}
\usage{
print_trials_js(trials, js_var = "trial_template")
}
\arguments{
\item{trials}{list returned by \code{trial_format()}}

\item{js_var}{JS variable name (default \code{"trial_template"})}
}
\description{
Print trial list as JavaScript
}
